requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.28
pydantic==2.6.1
pydantic-settings==2.2.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
requests==2.31.0
email-validator